librbd/cache/pwl: add space manager based on libpmem Library
RWL needs to implement replica in the future, so it needs its own space
allocator or manager to copy data directly to the remote server address
space during RDMA connection. After the pmem is mapped to the memory,
the cache of the local client and the remote client(replica) is only the
difference of the head address, and the others are replication. In this
way, the remote can reuse the current code. If the remote server uses a
allocator similar to libpmemobj to manage another set of processes, the
program will be very complex and the performance will be low. So we
decide abandon libpmemobj and use libpmem to manage the address space.

The new pmem manager provides the functions of device creation and
destruction, space allocation and release. Because pmem can be written
overlay, space allocation and release are only coordinate movement. In
terms of transaction, because pmem can only guarantee the atomicity of 8
bytes. double root and CRC are used to ensure that there is an valid
root in case of unexpected crash or downtime.

Signed-off-by: Yin Congmin <congmin.yin@intel.com>