cls/rbd: add rwlcache primaryping op.
This fix the following problem:
If daemon can't access osd cluster but the rdma-access is ok.
daemonping timedout and daemon removed. The primary still use the
daemon do replciated but the replicated data can't flush to osd if
primary crash. So we use priamryping to tell primary which has daemon
removed and primary should restart a new replciated-cache.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>