librbd/cache/pwl/ssd: fix bdev->aio_write() assert
If the given cache capacity is not aligned with the default block size(4k)
of bdev, the size of bdev will be smaller than the configured size, resulting
in aio_ Write() assert. When it detects whether the IO is valid, the offset
will be greater than the total bdev capacity after alignment.

Signed-off-by: Yin Congmin <congmin.yin@intel.com>