librbd/pwl/rwl/replica: Add replica server side implementation
Provide data redundancy function for RWL to avoid data loss
caused by RWL stand-alone failure.

Use RDMA(rpma library) to communicate between RWL and Replica.
a) Use the rpma send/recv function to complete the resource
   initialization and release of the Replica by RWL.
b) Using the rpma write function to complete the mirror copy
   of RWL. The writes requires zero CPU cycles on the Replica
   node.

DaemonPing is used for objclass to detect the health of Replica.
PrimaryPing is used for objclass to detect the health of RWL.

Signed-off-by: Feng Hualong <hualong.feng@intel.com>